# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  conversationId?: number | null;
  isPublished: boolean;
  /** @nullable */
  githubRepo?: string | null;
  /** @nullable */
  supabaseUrl?: string | null;
  /** @nullable */
  stripePublishableKey?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  name: string;
  description?: string;
}

export interface ProjectUpdate {
  name?: string;
  description?: string;
  /** @nullable */
  conversationId?: number | null;
}

export interface PublishProjectInput {
  isPublished: boolean;
}

export interface ProjectSettingsInput {
  /** @nullable */
  githubRepo?: string | null;
  /** @nullable */
  githubToken?: string | null;
  /** @nullable */
  supabaseUrl?: string | null;
  /** @nullable */
  supabaseAnonKey?: string | null;
  /** @nullable */
  stripePublishableKey?: string | null;
}

export interface GithubPushResult {
  success: boolean;
  repoUrl?: string;
  filesUpdated: number;
  message: string;
}

export interface ProjectFile {
  id: number;
  projectId: number;
  filename: string;
  content: string;
  language: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectFileInput {
  filename: string;
  content: string;
  language: string;
}

export interface ProjectFileUpdate {
  filename?: string;
  content?: string;
  language?: string;
}

export interface ProjectWithFiles {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  conversationId?: number | null;
  isPublished: boolean;
  /** @nullable */
  githubRepo?: string | null;
  /** @nullable */
  supabaseUrl?: string | null;
  /** @nullable */
  stripePublishableKey?: string | null;
  createdAt: string;
  updatedAt: string;
  files: ProjectFile[];
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}