# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project management
  - name: anthropic
    description: Anthropic AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with its files
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project with files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWithFiles"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /projects/{id}/files:
    get:
      operationId: listProjectFiles
      tags: [projects]
      summary: List files in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"
    post:
      operationId: createProjectFile
      tags: [projects]
      summary: Create or update a file in a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileInput"
      responses:
        "201":
          description: Created file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

  /projects/{id}/files/{fileId}:
    patch:
      operationId: updateProjectFile
      tags: [projects]
      summary: Update a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectFileUpdate"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteProjectFile
      tags: [projects]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/publish:
    post:
      operationId: publishProject
      tags: [projects]
      summary: Publish or unpublish a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /projects/{id}/settings:
    patch:
      operationId: updateProjectSettings
      tags: [projects]
      summary: Update project integration settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectSettingsInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /projects/{id}/github/push:
    post:
      operationId: pushProjectToGithub
      tags: [projects]
      summary: Push project files to a GitHub repository
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Push result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubPushResult"
        "400":
          description: Bad request (missing GitHub settings)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnthropicConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"

  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"

  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnthropicMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        conversationId:
          type: ["integer", "null"]
        isPublished:
          type: boolean
        githubRepo:
          type: ["string", "null"]
        supabaseUrl:
          type: ["string", "null"]
        stripePublishableKey:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isPublished
        - createdAt
        - updatedAt

    ProjectInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        conversationId:
          type: ["integer", "null"]

    PublishProjectInput:
      type: object
      properties:
        isPublished:
          type: boolean
      required:
        - isPublished

    ProjectSettingsInput:
      type: object
      properties:
        githubRepo:
          type: ["string", "null"]
        githubToken:
          type: ["string", "null"]
        supabaseUrl:
          type: ["string", "null"]
        supabaseAnonKey:
          type: ["string", "null"]
        stripePublishableKey:
          type: ["string", "null"]

    GithubPushResult:
      type: object
      properties:
        success:
          type: boolean
        repoUrl:
          type: string
        filesUpdated:
          type: integer
        message:
          type: string
      required:
        - success
        - filesUpdated
        - message

    ProjectFile:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        filename:
          type: string
        content:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - filename
        - content
        - language
        - createdAt
        - updatedAt

    ProjectFileInput:
      type: object
      properties:
        filename:
          type: string
        content:
          type: string
        language:
          type: string
      required:
        - filename
        - content
        - language

    ProjectFileUpdate:
      type: object
      properties:
        filename:
          type: string
        content:
          type: string
        language:
          type: string

    ProjectWithFiles:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        conversationId:
          type: ["integer", "null"]
        isPublished:
          type: boolean
        githubRepo:
          type: ["string", "null"]
        supabaseUrl:
          type: ["string", "null"]
        stripePublishableKey:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        files:
          type: array
          items:
            $ref: "#/components/schemas/ProjectFile"
      required:
        - id
        - name
        - isPublished
        - createdAt
        - updatedAt
        - files

    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    AnthropicConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    AnthropicMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error