# lib/api-zod/src/generated/types/projectSettingsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectSettingsInput {
  /** @nullable */
  githubRepo?: string | null;
  /** @nullable */
  githubToken?: string | null;
  /** @nullable */
  supabaseUrl?: string | null;
  /** @nullable */
  supabaseAnonKey?: string | null;
  /** @nullable */
  stripePublishableKey?: string | null;
}