# lib/api-zod/src/generated/types/projectWithFiles.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectFile } from './projectFile';

export interface ProjectWithFiles {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  conversationId?: number | null;
  isPublished: boolean;
  /** @nullable */
  githubRepo?: string | null;
  /** @nullable */
  supabaseUrl?: string | null;
  /** @nullable */
  stripePublishableKey?: string | null;
  createdAt: Date;
  updatedAt: Date;
  files: ProjectFile[];
}